move benchmark server code to its own file
